+++ app/sprinkles/core/src/Model/Relations/MorphManySyncable.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * UserFrosting (http://www.userfrosting.com)
+ *
+ * @link      https://github.com/userfrosting/UserFrosting
+ * @copyright Copyright (c) 2013-2017 Alexander Weissman
+ * @license   https://github.com/userfrosting/UserFrosting/blob/master/licenses/UserFrosting.md (MIT License)
+ */
+namespace UserFrosting\Sprinkle\Core\Model\Relations;
+
+use Illuminate\Database\Eloquent\Relations\MorphMany;
+use UserFrosting\Sprinkle\Core\Model\Relations\Concerns\Syncable;
+
+/**
+ * @link https://github.com/laravel/framework/blob/5.4/src/Illuminate/Database/Eloquent/Relations/MorphMany.php
+ */
+class MorphManySyncable extends MorphMany
+{
+    use Syncable;
+}