+++ app/sprinkles/core/src/Model/UFModel.php
@@ -3,13 +3,18 @@
  * UserFrosting (http://www.userfrosting.com)
  *
  * @link      https://github.com/userfrosting/UserFrosting
- * @copyright Copyright (c) 2013-2016 Alexander Weissman
+ * @copyright Copyright (c) 2013-2017 Alexander Weissman
  * @license   https://github.com/userfrosting/UserFrosting/blob/master/licenses/UserFrosting.md (MIT License)
  */
 namespace UserFrosting\Sprinkle\Core\Model;
 
 use Illuminate\Database\Capsule\Manager as Capsule;
 use Illuminate\Database\Eloquent\Model;
+use UserFrosting\Sprinkle\Core\Model\Relations\BelongsToManyConstrained;
+use UserFrosting\Sprinkle\Core\Model\Relations\BelongsToManyThrough;
+use UserFrosting\Sprinkle\Core\Model\Relations\BelongsToManyUnique;
+use UserFrosting\Sprinkle\Core\Model\Relations\HasManySyncable;
+use UserFrosting\Sprinkle\Core\Model\Relations\MorphManySyncable;
 
 /**
  * UFModel Class
@@ -50,6 +55,156 @@ abstract class UFModel extends Model
     }
 
     /**
+     * Define a constrained many-to-many relationship.
+     * This is similar to a regular many-to-many, but constrains the child results to match an additional constraint key in the parent object.
+     *
+     * @param  string  $related
+     * @param  string  $constraintKey
+     * @param  string  $table
+     * @param  string  $foreignKey
+     * @param  string  $relatedKey
+     * @param  string  $relation
+     * @return \UserFrosting\Sprinkle\Core\Model\Relations\BelongsToManyConstrained
+     */
+    public function belongsToManyConstrained($related, $constraintKey, $table = null, $foreignKey = null, $relatedKey = null, $relation = null)
+    {
+        // If no relationship name was passed, we will pull backtraces to get the
+        // name of the calling function. We will use that function name as the
+        // title of this relation since that is a great convention to apply.
+        if (is_null($relation)) {
+            $relation = $this->guessBelongsToManyRelation();
+        }
+
+        // First, we'll need to determine the foreign key and "other key" for the
+        // relationship. Once we have determined the keys we'll make the query
+        // instances as well as the relationship instances we need for this.
+        $instance = $this->newRelatedInstance($related);
+
+        $foreignKey = $foreignKey ?: $this->getForeignKey();
+
+        $relatedKey = $relatedKey ?: $instance->getForeignKey();
+
+        // If no table name was provided, we can guess it by concatenating the two
+        // models using underscores in alphabetical order. The two model names
+        // are transformed to snake case from their default CamelCase also.
+        if (is_null($table)) {
+            $table = $this->joiningTable($related);
+        }
+
+        return new BelongsToManyConstrained(
+            $instance->newQuery(), $this, $constraintKey, $table, $foreignKey, $relatedKey, $relation
+        );
+    }
+
+    /**
+     * Define a many-to-many 'through' relationship.
+     * This is basically hasManyThrough for many-to-many relationships.
+     *
+     * @param  string  $related
+     * @param  string  $through
+     * @param  string  $firstJoiningTable
+     * @param  string  $firstForeignKey
+     * @param  string  $firstRelatedKey
+     * @param  string  $secondJoiningTable
+     * @param  string  $secondForeignKey
+     * @param  string  $secondRelatedKey
+     * @param  string  $throughRelation
+     * @param  string  $relation
+     * @return \UserFrosting\Sprinkle\Core\Model\Relations\BelongsToManyThrough
+     */
+    public function belongsToManyThrough(
+        $related,
+        $through,
+        $firstJoiningTable = null,
+        $firstForeignKey = null,
+        $firstRelatedKey = null,
+        $secondJoiningTable = null,
+        $secondForeignKey = null,
+        $secondRelatedKey = null,
+        $throughRelation = null,
+        $relation = null
+    )
+    {
+        // If no relationship name was passed, we will pull backtraces to get the
+        // name of the calling function. We will use that function name as the
+        // title of this relation since that is a great convention to apply.
+        if (is_null($relation)) {
+            $relation = $this->guessBelongsToManyRelation();
+        }
+
+        // Create models for through and related
+        $through = new $through;
+        $related = $this->newRelatedInstance($related);
+
+        if (is_null($throughRelation)) {
+            $throughRelation = $through->getTable();
+        }
+
+        // If no table names were provided, we can guess it by concatenating the parent
+        // and through table names. The two model names are transformed to snake case
+        // from their default CamelCase also.
+        if (is_null($firstJoiningTable)) {
+            $firstJoiningTable = $this->joiningTable($through);
+        }
+
+        if (is_null($secondJoiningTable)) {
+            $secondJoiningTable = $through->joiningTable($related);
+        }
+
+        $firstForeignKey = $firstForeignKey ?: $this->getForeignKey();
+        $firstRelatedKey = $firstRelatedKey ?: $through->getForeignKey();
+        $secondForeignKey = $secondForeignKey ?: $through->getForeignKey();
+        $secondRelatedKey = $secondRelatedKey ?: $related->getForeignKey();
+
+        // This relationship maps the top model (this) to the through model.
+        $intermediateRelationship = $this->belongsToMany($through, $firstJoiningTable, $firstForeignKey, $firstRelatedKey, $throughRelation)
+            ->withPivot($firstForeignKey);
+
+        // Now we set up the relationship with the related model.
+        $query = new BelongsToManyThrough(
+            $related->newQuery(), $this, $intermediateRelationship, $secondJoiningTable, $secondForeignKey, $secondRelatedKey, $relation
+        );
+
+        return $query;
+    }
+
+    /**
+     * Define a unique many-to-many relationship.  Similar to a regular many-to-many relationship, but removes duplicate child objects.
+     *
+     * {@inheritDoc}
+     * @return \UserFrosting\Sprinkle\Core\Model\Relations\BelongsToManyUnique
+     */
+    public function belongsToManyUnique($related, $table = null, $foreignKey = null, $relatedKey = null, $relation = null)
+    {
+        // If no relationship name was passed, we will pull backtraces to get the
+        // name of the calling function. We will use that function name as the
+        // title of this relation since that is a great convention to apply.
+        if (is_null($relation)) {
+            $relation = $this->guessBelongsToManyRelation();
+        }
+
+        // First, we'll need to determine the foreign key and "other key" for the
+        // relationship. Once we have determined the keys we'll make the query
+        // instances as well as the relationship instances we need for this.
+        $instance = $this->newRelatedInstance($related);
+
+        $foreignKey = $foreignKey ?: $this->getForeignKey();
+
+        $relatedKey = $relatedKey ?: $instance->getForeignKey();
+
+        // If no table name was provided, we can guess it by concatenating the two
+        // models using underscores in alphabetical order. The two model names
+        // are transformed to snake case from their default CamelCase also.
+        if (is_null($table)) {
+            $table = $this->joiningTable($related);
+        }
+
+        return new BelongsToManyUnique(
+            $instance->newQuery(), $this, $table, $foreignKey, $relatedKey, $relation
+        );
+    }
+
+    /**
      * Get the properties of this object as an associative array.  Alias for toArray().
      *
      * @return array
@@ -57,6 +212,45 @@ abstract class UFModel extends Model
     public function export()
     {
         return $this->toArray();
+    }
+
+    /**
+     * Overrides the default Eloquent hasMany relationship to return a HasManySyncable.
+     *
+     * {@inheritDoc}
+     * @return \UserFrosting\Sprinkle\Core\Model\Relations\HasManySyncable
+     */
+    public function hasMany($related, $foreignKey = null, $localKey = null)
+    {
+        $instance = $this->newRelatedInstance($related);
+
+        $foreignKey = $foreignKey ?: $this->getForeignKey();
+
+        $localKey = $localKey ?: $this->getKeyName();
+
+        return new HasManySyncable(
+            $instance->newQuery(), $this, $instance->getTable().'.'.$foreignKey, $localKey
+        );
+    }
+
+    /**
+     * Overrides the default Eloquent morphMany relationship to return a MorphManySyncable.
+     *
+     * {@inheritDoc}
+     * @return \UserFrosting\Sprinkle\Core\Model\Relations\MorphManySyncable
+     */
+    public function morphMany($related, $name, $type = null, $id = null, $localKey = null)
+    {
+        $instance = $this->newRelatedInstance($related);
+
+        // Here we will gather up the morph type and ID for the relationship so that we
+        // can properly query the intermediate table of a relation. Finally, we will
+        // get the table and create the relationship instances for the developers.
+        list($type, $id) = $this->getMorphs($name, $type, $id);
+        $table = $instance->getTable();
+        $localKey = $localKey ?: $this->getKeyName();
+
+        return new MorphManySyncable($instance->newQuery(), $this, $table.'.'.$type, $table.'.'.$id, $localKey);
     }
 
     /**