+++ app/sprinkles/core/src/ServicesProvider/ServicesProvider.php
@@ -286,6 +286,23 @@ class ServicesProvider
             // Start Eloquent
             $capsule->bootEloquent();
 
+            if ($config['debug.queries']) {
+                $logger = $c->queryLogger;
+
+                foreach ($config['db'] as $name => $dbConfig) {
+                    $capsule->connection($name)->enableQueryLog();
+
+                    // Register listener
+                    $capsule->connection($name)->listen(function ($query) use ($logger) {
+                        $logger->debug("Query executed on database [{$query->connectionName}]:", [
+                            'query'    => $query->sql,
+                            'bindings' => $query->bindings,
+                            'time'     => $query->time . ' ms'
+                        ]);
+                    });
+                }
+            }
+
             return $capsule;
         };
 
@@ -432,6 +449,26 @@ class ServicesProvider
         };
 
         /**
+         * Laravel query logging with Monolog.
+         *
+         * Extend this service to push additional handlers onto the 'query' log stack.
+         */
+        $container['queryLogger'] = function ($c) {
+            $logger = new Logger('query');
+
+            $logFile = $c->locator->findResource('log://queries.log', true, true);
+
+            $handler = new StreamHandler($logFile);
+
+            $formatter = new MixedFormatter(null, null, true);
+
+            $handler->setFormatter($formatter);
+            $logger->pushHandler($handler);
+
+            return $logger;
+        };
+
+        /**
          * Override Slim's default router with the UF router.
          */
         $container['router'] = function ($c) {
@@ -561,6 +598,7 @@ class ServicesProvider
 
             if ($c->config['debug.twig']) {
                 $twig->enableDebug();
+                $view->addExtension(new \Twig_Extension_Debug());
             }
 
             // Register the Slim extension with Twig